Exporta card da Arte para componente ArtCard
